+++ .env.example
@@ -1,3 +1,4 @@
 BOT_NAME=
 BOT_AUTH=
 CHANNELS=
+PORT=
