+++ package.json
@@ -7,14 +7,18 @@
     "compile": "./node_modules/.bin/tsc --watch",
     "prestart": "tsc",
     "start": "node build/server.js",
-    "dev": "nodemon"
+    "dev": "nodemon",
+    "heroku-prebuild": "export NPM_CONFIG_PRODUCTION=false; export NODE_ENV=; NPM_CONFIG_PRODUCTION=false NODE_ENV=development npm install --only=dev --dev",
+    "heroku-postbuild": "export NPM_CONFIG_PRODUCTION=true; export NODE_ENV=production;"
   },
   "author": "whitep4nth3r",
   "license": "MIT",
   "dependencies": {
+    "dotenv": "^8.2.0",
     "express": "^4.17.1",
-    "ws": "^7.3.1",
-    "typescript": "^4.0.2"
+    "tmi.js": "^1.5.0",
+    "typescript": "^4.0.2",
+    "ws": "^7.3.1"
   },
   "devDependencies": {
     "@types/express": "^4.17.7",
